// apps/backend/src/common/error.types.ts
export enum HttpCode {
  NOT_FOUND = 404,
  BAD_REQUEST = 400,
  INTERNAL_SERVER_ERROR = 500,
  FORBIDDEN = 403,
  UNAUTHORIZED = 401
}

export enum CommonHttpError {
  NOT_FOUND = 'Not found',
  BAD_REQUEST = 'Bad request',
  INTERNAL_SERVER_ERROR = 'Internal server error',
  FORBIDDEN = 'Forbidden',
  UNAUTHORIZED = 'Unathorized'
}